Sniper added, changes weapon damage types to integers, 'reworked' weapon switching cooldowns

+++ Assets/Scripts/Weapon.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public abstract class Weapon : MonoBehaviour, ILoadScriptable
 {
     [HideInInspector] public int energyCost;
-    [HideInInspector] public float damage;
+    [HideInInspector] public int damage;
     [HideInInspector] public float firingSpeed;
     [HideInInspector] public Sprite weaponSprite;
     [HideInInspector] public GameObject projectile;
